pyproject.toml
[project]
name = "psych-articles"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "polars>=1.36.1",
    "pyalex>=0.19",
    "pyarrow>=22.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "semanticscholar>=0.11.0",
    "webdataset>=1.0.2",
]
